% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{angle_unit}
\alias{angle_unit}
\title{The angular unit a frame declares}
\usage{
angle_unit(data)
}
\arguments{
\item{data}{An aniframe.}
}
\value{
\code{"deg"} or \code{"rad"}. Anything other than degrees is read as radians,
which is what the angular computations use internally.
}
\description{
The angular unit a frame declares
}
\keyword{internal}
