% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_direction.R
\name{calculate_direction_2d}
\alias{calculate_direction_2d}
\title{Calculate direction
Calculate direction (angle) from x and y distance using the (two-argument) arc-tangent.}
\usage{
calculate_direction_2d(dx, dy)
}
\arguments{
\item{dx}{dx}

\item{dy}{dy}
}
\description{
Calculate direction
Calculate direction (angle) from x and y distance using the (two-argument) arc-tangent.
}
\keyword{internal}
