% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_direction.R
\name{calculate_direction_3d}
\alias{calculate_direction_3d}
\title{Calculate direction
Calculate direction (angle) from x, y and z distance using ...}
\usage{
calculate_direction_3d(dx, dy, dz)
}
\arguments{
\item{dx}{dx}

\item{dy}{dy}

\item{dz}{dz}
}
\description{
Calculate direction
Calculate direction (angle) from x, y and z distance using ...
}
\keyword{internal}
