% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_kinematics_2d}
\alias{calculate_kinematics_2d}
\title{Calculate kinematics from 2D movements}
\usage{
calculate_kinematics_2d(data)
}
\arguments{
\item{data}{An aniframe}
}
\value{
An aniframe with the original columns plus the kinematic measures:
\code{distance}, \code{cumsum_distance}, \code{v_translation}, \code{a_translation},
\code{direction}, \code{rotation}, \code{cumsum_rotation}, \code{v_rotation}, \code{a_rotation}.
}
\description{
Calculate kinematics from 2D movements
}
