% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_kinematics_3d}
\alias{calculate_kinematics_3d}
\title{Calculate kinematics from 3D movements}
\usage{
calculate_kinematics_3d(data)
}
\arguments{
\item{data}{An aniframe}
}
\value{
An aniframe with the original columns plus the kinematic measures:
\code{distance}, \code{cumsum_distance}, \code{v_translation}, \code{a_translation},
}
\description{
Calculate kinematics from 3D movements
}
