% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_kinematics.R
\name{calculate_summaries}
\alias{calculate_summaries}
\title{Calculate summary statistics for kinematic measures}
\usage{
calculate_summaries(data, measures, is_3d = FALSE)
}
\description{
Calculate summary statistics for kinematic measures
}
\keyword{internal}
