% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_kinematics.R
\name{calculate_totals}
\alias{calculate_totals}
\title{Calculate totals and tortuosity metrics}
\usage{
calculate_totals(data, is_3d = FALSE)
}
\description{
Calculate totals and tortuosity metrics
}
\keyword{internal}
