% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_translation_2d}
\alias{calculate_translation_2d}
\title{Calculate translational kinematics in 2D}
\usage{
calculate_translation_2d(data)
}
\arguments{
\item{data}{An anipoint with x, y, and time columns}
}
\value{
The anipoint with added translational kinematic columns
}
\description{
Calculate translational kinematics in 2D
}
\keyword{internal}
