% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_orientation.R
\name{check_members}
\alias{check_members}
\title{Check that arguments name members of a level}
\usage{
check_members(
  x,
  members,
  level,
  arg,
  single = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{The argument's value.}

\item{members}{The level's members.}

\item{level}{The level's name, for the message.}

\item{arg}{The argument's name, for the message.}

\item{single}{Whether exactly one member is expected.}

\item{call}{The calling environment, for error messages.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Check that arguments name members of a level
}
\keyword{internal}
