% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kinematics.R
\name{check_min_step}
\alias{check_min_step}
\title{Check a \code{min_step} argument}
\usage{
check_min_step(min_step, call = rlang::caller_env())
}
\arguments{
\item{min_step}{See \code{\link[=add_kinematics]{add_kinematics()}}.}

\item{call}{The calling environment, for error messages.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Check a \code{min_step} argument
}
\keyword{internal}
