% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rediscretise.R
\name{check_step_length}
\alias{check_step_length}
\title{Check a \code{step_length} argument}
\usage{
check_step_length(step_length, call = rlang::caller_env())
}
\arguments{
\item{step_length}{See \code{\link[=summarise_path]{summarise_path()}}.}

\item{call}{The calling environment, for error messages.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Check a \code{step_length} argument
}
\keyword{internal}
