% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_aniframe.R
\name{check_summary_cols}
\alias{check_summary_cols}
\title{Check the columns asked to be summarised}
\usage{
check_summary_cols(data, cols, call = rlang::caller_env())
}
\arguments{
\item{data}{An aniframe.}

\item{cols}{Character vector of column names.}

\item{call}{The calling environment, for error messages.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Check the columns asked to be summarised
}
\keyword{internal}
