% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tortuosity.R
\name{check_tortuosity_input}
\alias{check_tortuosity_input}
\title{Check the input to the windowed tortuosity}
\usage{
check_tortuosity_input(data, window_width, call = rlang::caller_env())
}
\arguments{
\item{data}{An anipoint.}

\item{window_width}{The window width asked for.}

\item{call}{The calling environment, for error messages.}
}
\value{
\code{window_width}, as an integer.
}
\description{
Check the input to the windowed tortuosity
}
\keyword{internal}
