% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_descriptive.R
\name{circ_mad}
\alias{circ_mad}
\title{Circular median absolute deviation}
\usage{
circ_mad(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles (radians)}

\item{na_rm}{Logical; if TRUE, remove NA values before computation}
}
\value{
MAD (radians)
}
\description{
Circular median absolute deviation
}
\keyword{internal}
