% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_descriptive.R
\name{circ_mean}
\alias{circ_mean}
\title{Circular mean}
\usage{
circ_mean(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles in radians}

\item{na_rm}{Logical; if TRUE, remove NA values before computation}
}
\value{
Mean direction (radians) wrapped to [0, 2*pi)
}
\description{
Circular mean
}
\keyword{internal}
