% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_descriptive.R
\name{circ_sd}
\alias{circ_sd}
\title{Circular standard deviation}
\usage{
circ_sd(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles (radians)}

\item{na_rm}{Logical; if TRUE, remove NA values before computation}
}
\value{
Circular SD (radians)
}
\description{
Circular standard deviation
}
\keyword{internal}
