% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{cross_rows}
\alias{cross_rows}
\title{Row-wise cross product of two three-column matrices}
\usage{
cross_rows(u, v)
}
\arguments{
\item{u, v}{Numeric matrices with three columns.}
}
\value{
A numeric matrix with three columns.
}
\description{
Row-wise cross product of two three-column matrices
}
\keyword{internal}
