% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniframe_kin_is.R
\name{ensure_is_aniframe_kin}
\alias{ensure_is_aniframe_kin}
\title{Ensure object is an aniframe_kin}
\usage{
ensure_is_aniframe_kin(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Error if not an aniframe_kin
}
\description{
Ensure object is an aniframe_kin
}
\keyword{internal}
