% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_aniframe.R
\name{join_summaries}
\alias{join_summaries}
\title{Join multiple summary data frames}
\usage{
join_summaries(summaries, group_vars)
}
\description{
Join multiple summary data frames
}
\keyword{internal}
