% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kinematics.R
\name{kinematics_cartesian}
\alias{kinematics_cartesian}
\title{Add translational, and where defined rotational, kinematics}
\usage{
kinematics_cartesian(data, vertical = NULL)
}
\arguments{
\item{data}{A Cartesian anipoint.}

\item{vertical}{See \code{\link[=add_kinematics]{add_kinematics()}}.}
}
\value{
The anipoint with added kinematic columns.
}
\description{
Add translational, and where defined rotational, kinematics
}
\keyword{internal}
