% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_angle.R
\name{mean_angle}
\alias{mean_angle}
\title{Compute the circular mean of angles}
\usage{
mean_angle(ang)
}
\arguments{
\item{ang}{Numeric vector of angles in radians.}
}
\value{
Circular mean in [0, 2*pi).
}
\description{
Returns the mean direction of a vector of angles in radians,
wrapped to [0, 2*pi).
}
\examples{
mean_angle(c(pi/2, 1.5 * pi))
mean_angle(c(0, pi))
}
