% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_angle.R
\name{median_angle}
\alias{median_angle}
\title{Compute the circular median of angles}
\usage{
median_angle(ang)
}
\arguments{
\item{ang}{Numeric vector of angles in radians.}
}
\value{
Circular median in [0, 2*pi).
}
\description{
Returns the median direction of a vector of angles in radians,
wrapped to [0, 2*pi).
}
\examples{
median_angle(c(pi/2, 1.5 * pi))
median_angle(c(0, pi))
}
