% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniframe_kin_new.R
\name{new_aniframe_kin}
\alias{new_aniframe_kin}
\title{Create a new aniframe_kin object (internal constructor)}
\usage{
new_aniframe_kin(x)
}
\arguments{
\item{x}{An anipoint to add the aniframe_kin class to}
}
\value{
An aniframe_kin object
}
\description{
Create a new aniframe_kin object (internal constructor)
}
\keyword{internal}
