% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rediscretise.R
\name{path_sinuosity}
\alias{path_sinuosity}
\title{Sinuosity and E_max of a rediscretised path}
\usage{
path_sinuosity(position, time)
}
\arguments{
\item{position}{A data frame of positions, one column per axis.}

\item{time}{The index.}
}
\value{
A list of \code{sinuosity} and \code{e_max}, each a number.
}
\description{
Sinuosity and E_max of a rediscretised path
}
\keyword{internal}
