% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rad_to_deg}
\alias{deg_to_rad}
\alias{wrap_angle}
\alias{unwrap_angle}
\alias{calculate_angular_difference}
\alias{diff_angle}
\alias{is_aniframe}
\alias{ensure_is_aniframe}
\alias{as_aniframe}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{aniframe}{\code{\link[anicore:as_aniframe]{as_aniframe()}}, \code{\link[anicore:deg_to_rad]{deg_to_rad()}}, \code{\link[anicore:ensure_is_aniframe]{ensure_is_aniframe()}}, \code{\link[anicore:is_aniframe]{is_aniframe()}}, \code{\link[anicore:rad_to_deg]{rad_to_deg()}}}

  \item{anispace}{\code{\link[anispace:calculate_angular_difference]{calculate_angular_difference()}}, \code{\link[anispace:diff_angle]{diff_angle()}}, \code{\link[anispace:unwrap_angle]{unwrap_angle()}}, \code{\link[anispace:wrap_angle]{wrap_angle()}}}
}}

