% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{add_kinematics}
\alias{add_kinematics}
\title{Add translational, and where defined rotational, kinematics}
\usage{
add_kinematics(data, vertical = NULL)
}
\arguments{
\item{data}{A Cartesian anipoint.}

\item{vertical}{See \code{\link[=calculate_kinematics]{calculate_kinematics()}}.}
}
\value{
The anipoint with added kinematic columns.
}
\description{
Add translational, and where defined rotational, kinematics
}
\keyword{internal}
