% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animetric-package.R
\docType{package}
\name{animetric-package}
\alias{animetric}
\alias{animetric-package}
\title{animetric: An R package for calculating movement-based metrics}
\description{
An R package for calculating movement-based metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://animovement.dev/animetric/}
  \item \url{https://github.com/animovement/animetric/}
  \item Report bugs at \url{https://github.com/animovement/animetric/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arbøl \email{animovement.84w1m@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

}
\keyword{internal}
