% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_rotation_2d}
\alias{calculate_rotation_2d}
\title{Calculate rotational kinematics in 2D}
\usage{
calculate_rotation_2d(data)
}
\arguments{
\item{data}{An aniframe with v_x, v_y, and time columns}
}
\value{
The aniframe with added rotational kinematic columns
}
\description{
Computes heading angles and angular kinematics based on the velocity vector.
Heading is calculated as atan2(v_y, v_x).
}
\keyword{internal}
