% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_translation_3d}
\alias{calculate_translation_3d}
\title{Calculate translational kinematics in 3D}
\usage{
calculate_translation_3d(data)
}
\arguments{
\item{data}{An aniframe with x, y, z, and time columns}
}
\value{
The aniframe with added translational kinematic columns
}
\description{
Calculate translational kinematics in 3D
}
\keyword{internal}
