% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{check_vertical}
\alias{check_vertical}
\title{Check a \code{vertical} argument}
\usage{
check_vertical(vertical, call = rlang::caller_env())
}
\arguments{
\item{vertical}{See \code{\link[=calculate_kinematics]{calculate_kinematics()}}.}

\item{call}{The calling environment, for error messages.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Check a \code{vertical} argument
}
\keyword{internal}
