% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kinematics.R
\name{horizontal_basis}
\alias{horizontal_basis}
\title{The horizontal axes course is measured in}
\usage{
horizontal_basis(up)
}
\arguments{
\item{up}{A unit vertical along one of the axes.}
}
\value{
A list of two unit vectors, \code{first} (course 0) and \code{second}
(course pi/2).
}
\description{
Course counts about the vertical by the right-hand rule, from the axis
after it in the cycle x, y, z.
}
\keyword{internal}
