% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniframe_kin_is.R
\name{is_aniframe_kin}
\alias{is_aniframe_kin}
\title{Check if object is an aniframe_kin}
\usage{
is_aniframe_kin(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical: TRUE if x inherits from aniframe
}
\description{
Check if object is an aniframe_kin
}
