% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rediscretise.R
\name{path_sinuosity}
\alias{path_sinuosity}
\title{Sinuosity and E_max of a rediscretised path}
\usage{
path_sinuosity(position, time, step_length = "auto")
}
\arguments{
\item{position}{A data frame of positions, one column per axis.}

\item{time}{The index.}

\item{step_length}{The step to rediscretise at, as for
\code{\link[=resolve_step_length]{resolve_step_length()}}.}
}
\value{
A list of \code{sinuosity} and \code{e_max}, each a number.
}
\description{
Sinuosity and E_max of a rediscretised path
}
\keyword{internal}
