% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-aniframe.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rad_to_deg}
\alias{deg_to_rad}
\alias{wrap_angle}
\alias{unwrap_angle}
\alias{calculate_angular_difference}
\alias{diff_angle}
\alias{is_aniframe}
\alias{ensure_is_aniframe}
\alias{as_aniframe}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{aniframe}{\code{\link[aniframe]{as_aniframe}}, \code{\link[aniframe]{calculate_angular_difference}}, \code{\link[aniframe]{deg_to_rad}}, \code{\link[aniframe]{diff_angle}}, \code{\link[aniframe]{ensure_is_aniframe}}, \code{\link[aniframe]{is_aniframe}}, \code{\link[aniframe]{rad_to_deg}}, \code{\link[aniframe]{unwrap_angle}}, \code{\link[aniframe]{wrap_angle}}}
}}

