% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{vertical_vector}
\alias{vertical_vector}
\title{The unit vertical of a frame}
\usage{
vertical_vector(n_axes, vertical)
}
\arguments{
\item{n_axes}{How many Cartesian axes the frame has.}

\item{vertical}{See \code{\link[=calculate_kinematics]{calculate_kinematics()}}, already checked.}
}
\value{
A length-3 unit vector, or \code{NULL} when a 3D frame has no
\code{vertical}. In 2D it is the normal of the data's plane.
}
\description{
The unit vertical of a frame
}
\keyword{internal}
